pyproject.toml
[tool.poetry]
name = "lucytok"
version = "0.1.2"
description = ""
authors = ["Doug Turnbull <softwaredoug@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pystemmer = "^2.2.0.3"
regex = "^2024.9.11"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
